+++ api/feature.go
@@ -8,6 +8,7 @@ const (
 	CoarseCurrent            // charger
 	IntegratedDevice         // charger
 	Heating                  // charger
+	Average                  // tariff
 	Cacheable                // tariff
 	Offline                  // vehicle
 	Retryable                // vehicle
