+++ tariff/entsoe.go
@@ -20,11 +20,10 @@ import (
 type Entsoe struct {
 	*request.Helper
 	*embed
-	log     *util.Logger
-	token   string
-	domain  string
-	average time.Duration
-	data    *util.Monitor[api.Rates]
+	log    *util.Logger
+	token  string
+	domain string
+	data   *util.Monitor[api.Rates]
 }
 
 var _ api.Tariff = (*Entsoe)(nil)
@@ -38,7 +37,6 @@ func NewEntsoeFromConfig(other map[string]interface{}) (api.Tariff, error) {
 		embed         `mapstructure:",squash"`
 		Securitytoken string
 		Domain        string
-		Average       time.Duration
 	}
 
 	if err := util.DecodeOther(other, &cc); err != nil {
@@ -65,13 +63,12 @@ func NewEntsoeFromConfig(other map[string]interface{}) (api.Tariff, error) {
 	log := util.NewLogger("entsoe").Redact(cc.Securitytoken)
 
 	t := &Entsoe{
-		log:     log,
-		Helper:  request.NewHelper(log),
-		embed:   &cc.embed,
-		token:   cc.Securitytoken,
-		domain:  domain,
-		average: cc.Average,
-		data:    util.NewMonitor[api.Rates](2 * time.Hour),
+		log:    log,
+		Helper: request.NewHelper(log),
+		embed:  &cc.embed,
+		token:  cc.Securitytoken,
+		domain: domain,
+		data:   util.NewMonitor[api.Rates](2 * time.Hour),
 	}
 
 	// Wrap the client with a decorator that adds the security token to each request.
@@ -169,10 +166,6 @@ func (t *Entsoe) Rates() (api.Rates, error) {
 	err := t.data.GetFunc(func(val api.Rates) {
 		res = slices.Clone(val)
 	})
-
-	if t.average > 0 {
-		res = averageSlots(res, t.average)
-	}
 
 	return res, err
 }
