+++ tariff/helper_test.go
@@ -86,28 +86,3 @@ func TestRunOrQError(t *testing.T) {
 		require.Nil(t, res)
 	}
 }
-
-func TestAverage(t *testing.T) {
-	clock := clock.NewMock()
-	clock.Add(30 * time.Minute)
-
-	var rr api.Rates
-	for i := range 5 {
-		rr = append(rr, api.Rate{
-			Start: clock.Now(),
-			End:   clock.Now().Add(SlotDuration),
-			Value: float64(i + 1),
-		})
-		clock.Add(SlotDuration)
-	}
-
-	res := averageSlots(rr, time.Hour)
-
-	rr[0].Value = 1.5
-	rr[1].Value = 1.5
-	rr[2].Value = 4.0
-	rr[3].Value = 4.0
-	rr[4].Value = 4.0
-
-	require.Equal(t, rr, res)
-}
