+++ util/templates/includes/tariff-features.tpl
@@ -1,0 +1,5 @@
+{{ define "tariff-features" }}
+{{- if eq .average "true" }}
+features: ["average"]
+{{- end }}
+{{- end }}
